backend/pyproject.toml
[project]
name = "text-heatmap-backend"
version = "0.1.0"
description = "LLM attention 기반 텍스트 중요도 분석 API"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "transformers>=4.37.0",
    "torch>=2.1.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.1",
    "python-docx>=1.1.0",
]

[tool.uv]
dev-dependencies = []
